LayoutTests/imported/w3c/web-platform-tests/selection/setBaseAndExtent-expected.txt


PASS Range 0 [paras[0].firstChild, 0, paras[0].firstChild, 0] setBaseAndExtent() 
PASS Reverse range 0 [paras[0].firstChild, 0, paras[0].firstChild, 0] setBaseAndExtent() 
FAIL Range 1 [paras[0].firstChild, 0, paras[0].firstChild, 1] setBaseAndExtent() assert_equals: focusOffset must equal the requested focus offset expected 1 but got 2
FAIL Reverse range 1 [paras[0].firstChild, 0, paras[0].firstChild, 1] setBaseAndExtent() assert_equals: anchorOffset must equal the requested focus offset expected 1 but got 2
PASS Range 2 [paras[0].firstChild, 2, paras[0].firstChild, 8] setBaseAndExtent() 
PASS Reverse range 2 [paras[0].firstChild, 2, paras[0].firstChild, 8] setBaseAndExtent() 
FAIL Range 3 [paras[0].firstChild, 2, paras[0].firstChild, 9] setBaseAndExtent() assert_equals: focusOffset must equal the requested focus offset expected 9 but got 10
FAIL Reverse range 3 [paras[0].firstChild, 2, paras[0].firstChild, 9] setBaseAndExtent() assert_equals: anchorOffset must equal the requested focus offset expected 9 but got 10
FAIL Range 4 [paras[1].firstChild, 0, paras[1].firstChild, 0] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Text node "Ijklmnop
" but got Text node "Qrstuvwx"
FAIL Reverse range 4 [paras[1].firstChild, 0, paras[1].firstChild, 0] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Text node "Ijklmnop
" but got Text node "Qrstuvwx"
FAIL Range 5 [paras[1].firstChild, 0, paras[1].firstChild, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Text node "Ijklmnop
" but got Text node "Qrstuvwx"
FAIL Reverse range 5 [paras[1].firstChild, 0, paras[1].firstChild, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Text node "Ijklmnop
" but got Text node "Qrstuvwx"
FAIL Range 6 [paras[1].firstChild, 2, paras[1].firstChild, 8] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Text node "Ijklmnop
" but got Text node "Qrstuvwx"
FAIL Reverse range 6 [paras[1].firstChild, 2, paras[1].firstChild, 8] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Text node "Ijklmnop
" but got Text node "Qrstuvwx"
FAIL Range 7 [paras[1].firstChild, 2, paras[1].firstChild, 9] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Text node "Ijklmnop
" but got Text node "Qrstuvwx"
FAIL Reverse range 7 [paras[1].firstChild, 2, paras[1].firstChild, 9] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Text node "Ijklmnop
" but got Text node "Qrstuvwx"
PASS Range 8 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 0] setBaseAndExtent() 
PASS Reverse range 8 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 0] setBaseAndExtent() 
PASS Range 9 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 1] setBaseAndExtent() 
PASS Reverse range 9 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 1] setBaseAndExtent() 
PASS Range 10 [detachedPara1.firstChild, 2, detachedPara1.firstChild, 8] setBaseAndExtent() 
PASS Reverse range 10 [detachedPara1.firstChild, 2, detachedPara1.firstChild, 8] setBaseAndExtent() 
PASS Range 11 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 0] setBaseAndExtent() 
PASS Reverse range 11 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 0] setBaseAndExtent() 
PASS Range 12 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 1] setBaseAndExtent() 
PASS Reverse range 12 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 1] setBaseAndExtent() 
PASS Range 13 [foreignPara1.firstChild, 2, foreignPara1.firstChild, 8] setBaseAndExtent() 
PASS Reverse range 13 [foreignPara1.firstChild, 2, foreignPara1.firstChild, 8] setBaseAndExtent() 
FAIL Range 14 [document.documentElement, 0, document.documentElement, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <html><head><title>Selection.setBaseAndExtent() tests</ti... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 14 [document.documentElement, 0, document.documentElement, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <html><head><title>Selection.setBaseAndExtent() tests</ti... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Range 15 [document.documentElement, 0, document.documentElement, 2] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <html><head><title>Selection.setBaseAndExtent() tests</ti... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 15 [document.documentElement, 0, document.documentElement, 2] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <html><head><title>Selection.setBaseAndExtent() tests</ti... but got Text node "Qrstuvwx"
FAIL Range 16 [document.documentElement, 1, document.documentElement, 2] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <html><head><title>Selection.setBaseAndExtent() tests</ti... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 16 [document.documentElement, 1, document.documentElement, 2] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <html><head><title>Selection.setBaseAndExtent() tests</ti... but got Text node "Qrstuvwx"
FAIL Range 17 [document.head, 1, document.head, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <head><title>Selection.setBaseAndExtent() tests</title>
<... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 17 [document.head, 1, document.head, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <head><title>Selection.setBaseAndExtent() tests</title>
<... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Range 18 [document.body, 0, document.body, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <body><div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 18 [document.body, 0, document.body, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <body><div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id... but got Text node "Qrstuvwx"
PASS Range 19 [foreignDoc.documentElement, 0, foreignDoc.documentElement, 1] setBaseAndExtent() 
PASS Reverse range 19 [foreignDoc.documentElement, 0, foreignDoc.documentElement, 1] setBaseAndExtent() 
PASS Range 20 [foreignDoc.head, 1, foreignDoc.head, 1] setBaseAndExtent() 
PASS Reverse range 20 [foreignDoc.head, 1, foreignDoc.head, 1] setBaseAndExtent() 
PASS Range 21 [foreignDoc.body, 0, foreignDoc.body, 0] setBaseAndExtent() 
PASS Reverse range 21 [foreignDoc.body, 0, foreignDoc.body, 0] setBaseAndExtent() 
FAIL Range 22 [paras[0], 0, paras[0], 0] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p> but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 22 [paras[0], 0, paras[0], 0] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p> but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Range 23 [paras[0], 0, paras[0], 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p> but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 23 [paras[0], 0, paras[0], 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p> but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
PASS Range 24 [detachedPara1, 0, detachedPara1, 0] setBaseAndExtent() 
PASS Reverse range 24 [detachedPara1, 0, detachedPara1, 0] setBaseAndExtent() 
PASS Range 25 [detachedPara1, 0, detachedPara1, 1] setBaseAndExtent() 
PASS Reverse range 25 [detachedPara1, 0, detachedPara1, 1] setBaseAndExtent() 
FAIL Range 26 [paras[0].firstChild, 0, paras[1].firstChild, 0] setBaseAndExtent() assert_equals: focusNode must equal the requested focus node expected Text node "Ijklmnop
" but got Element node <p id="c">Qrstuvwx</p>
FAIL Reverse range 26 [paras[0].firstChild, 0, paras[1].firstChild, 0] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Text node "Ijklmnop
" but got Element node <p id="c">Qrstuvwx</p>
FAIL Range 27 [paras[0].firstChild, 0, paras[1].firstChild, 8] setBaseAndExtent() assert_equals: focusNode must equal the requested focus node expected Text node "Ijklmnop
" but got Element node <p id="c">Qrstuvwx</p>
FAIL Reverse range 27 [paras[0].firstChild, 0, paras[1].firstChild, 8] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Text node "Ijklmnop
" but got Element node <p id="c">Qrstuvwx</p>
FAIL Range 28 [paras[0].firstChild, 3, paras[3], 1] setBaseAndExtent() assert_equals: anchorOffset must equal the requested anchor offset expected 3 but got 4
FAIL Reverse range 28 [paras[0].firstChild, 3, paras[3], 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL Range 29 [paras[0], 0, paras[0].firstChild, 7] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p> but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 29 [paras[0], 0, paras[0].firstChild, 7] setBaseAndExtent() assert_equals: anchorOffset must equal the requested focus offset expected 7 but got 8
FAIL Range 30 [testDiv, 2, paras[4], 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s... but got Text node "Qrstuvwx"
FAIL Reverse range 30 [testDiv, 2, paras[4], 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <p id="e" style="display:none">Ghijklmn</p> but got Text node "Qrstuvwx"
FAIL Range 31 [testDiv, 1, paras[2].firstChild, 5] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s... but got Text node "Qrstuvwx"
FAIL Reverse range 31 [testDiv, 1, paras[2].firstChild, 5] setBaseAndExtent() assert_equals: focusNode must equal the requested anchor node expected Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s... but got Text node "Qrstuvwx"
FAIL Range 32 [document.documentElement, 1, document.body, 0] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <html><head><title>Selection.setBaseAndExtent() tests</ti... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 32 [document.documentElement, 1, document.body, 0] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Element node <body><div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
PASS Range 33 [foreignDoc.documentElement, 1, foreignDoc.body, 0] setBaseAndExtent() 
PASS Reverse range 33 [foreignDoc.documentElement, 1, foreignDoc.body, 0] setBaseAndExtent() 
FAIL Range 34 [document, 0, document, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Document node with 2 children but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 34 [document, 0, document, 1] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Document node with 2 children but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Range 35 [document, 0, document, 2] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Document node with 2 children but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 35 [document, 0, document, 2] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Document node with 2 children but got Text node "Qrstuvwx"
FAIL Range 36 [document, 1, document, 2] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Document node with 2 children but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 36 [document, 1, document, 2] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Document node with 2 children but got Text node "Qrstuvwx"
FAIL Range 37 [testDiv, 0, comment, 5] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <div id="test"><p id="a">Äb̈c̈d̈ëf̈g̈ḧ
</p><p id="b" s... but got Text node "Äb̈c̈d̈ëf̈g̈ḧ
"
FAIL Reverse range 37 [testDiv, 0, comment, 5] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Comment node <!--Alphabet soup?--> but got Text node "Qrstuvwx"
FAIL Range 38 [paras[2].firstChild, 4, comment, 2] setBaseAndExtent() assert_equals: focusNode must equal the requested focus node expected Comment node <!--Alphabet soup?--> but got Text node "Qrstuvwx"
FAIL Reverse range 38 [paras[2].firstChild, 4, comment, 2] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Comment node <!--Alphabet soup?--> but got Text node "Qrstuvwx"
FAIL Range 39 [paras[3], 1, comment, 8] setBaseAndExtent() assert_equals: anchorNode must equal the requested anchor node expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL Reverse range 39 [paras[3], 1, comment, 8] setBaseAndExtent() assert_equals: anchorNode must equal the requested focus node expected Comment node <!--Alphabet soup?--> but got Text node "Qrstuvwx"
PASS Range 40 [foreignDoc, 0, foreignDoc, 0] setBaseAndExtent() 
PASS Reverse range 40 [foreignDoc, 0, foreignDoc, 0] setBaseAndExtent() 
PASS Range 41 [foreignDoc, 1, foreignComment, 2] setBaseAndExtent() 
PASS Reverse range 41 [foreignDoc, 1, foreignComment, 2] setBaseAndExtent() 
PASS Range 42 [foreignDoc.body, 0, foreignTextNode, 36] setBaseAndExtent() 
PASS Reverse range 42 [foreignDoc.body, 0, foreignTextNode, 36] setBaseAndExtent() 
PASS Range 43 [xmlDoc, 0, xmlDoc, 0] setBaseAndExtent() 
PASS Reverse range 43 [xmlDoc, 0, xmlDoc, 0] setBaseAndExtent() 
PASS Range 44 [xmlDoc, 1, xmlComment, 0] setBaseAndExtent() 
PASS Reverse range 44 [xmlDoc, 1, xmlComment, 0] setBaseAndExtent() 
PASS Range 45 [detachedTextNode, 0, detachedTextNode, 8] setBaseAndExtent() 
PASS Reverse range 45 [detachedTextNode, 0, detachedTextNode, 8] setBaseAndExtent() 
PASS Range 46 [detachedForeignTextNode, 7, detachedForeignTextNode, 7] setBaseAndExtent() 
PASS Reverse range 46 [detachedForeignTextNode, 7, detachedForeignTextNode, 7] setBaseAndExtent() 
PASS Range 47 [detachedForeignTextNode, 0, detachedForeignTextNode, 8] setBaseAndExtent() 
PASS Reverse range 47 [detachedForeignTextNode, 0, detachedForeignTextNode, 8] setBaseAndExtent() 
PASS Range 48 [detachedXmlTextNode, 7, detachedXmlTextNode, 7] setBaseAndExtent() 
PASS Reverse range 48 [detachedXmlTextNode, 7, detachedXmlTextNode, 7] setBaseAndExtent() 
PASS Range 49 [detachedXmlTextNode, 0, detachedXmlTextNode, 8] setBaseAndExtent() 
PASS Reverse range 49 [detachedXmlTextNode, 0, detachedXmlTextNode, 8] setBaseAndExtent() 
PASS Range 50 [detachedComment, 3, detachedComment, 4] setBaseAndExtent() 
PASS Reverse range 50 [detachedComment, 3, detachedComment, 4] setBaseAndExtent() 
PASS Range 51 [detachedComment, 5, detachedComment, 5] setBaseAndExtent() 
PASS Reverse range 51 [detachedComment, 5, detachedComment, 5] setBaseAndExtent() 
PASS Range 52 [detachedForeignComment, 0, detachedForeignComment, 1] setBaseAndExtent() 
PASS Reverse range 52 [detachedForeignComment, 0, detachedForeignComment, 1] setBaseAndExtent() 
PASS Range 53 [detachedForeignComment, 4, detachedForeignComment, 4] setBaseAndExtent() 
PASS Reverse range 53 [detachedForeignComment, 4, detachedForeignComment, 4] setBaseAndExtent() 
PASS Range 54 [detachedXmlComment, 2, detachedXmlComment, 6] setBaseAndExtent() 
PASS Reverse range 54 [detachedXmlComment, 2, detachedXmlComment, 6] setBaseAndExtent() 
PASS Range 55 [docfrag, 0, docfrag, 0] setBaseAndExtent() 
PASS Reverse range 55 [docfrag, 0, docfrag, 0] setBaseAndExtent() 
PASS Range 56 [foreignDocfrag, 0, foreignDocfrag, 0] setBaseAndExtent() 
PASS Reverse range 56 [foreignDocfrag, 0, foreignDocfrag, 0] setBaseAndExtent() 
PASS Range 57 [xmlDocfrag, 0, xmlDocfrag, 0] setBaseAndExtent() 
PASS Reverse range 57 [xmlDocfrag, 0, xmlDocfrag, 0] setBaseAndExtent() 
FAIL setBaseAndExtent() with index larger than length assert_equals: focus offset, got an IndexSizeError exception expected "IndexSizeError" but got "Error"
FAIL setBaseAndExtent() with negative index assert_equals: focus offset, got an IndexSizeError exception expected "IndexSizeError" but got "Error"
FAIL setBaseAndExtent() with null nodes assert_equals: focus node, got an TypeError exception expected "TypeError" but got "Error"
PASS setBaseAndExtent() with too few params 
